Add testapp support with pattern definitions and utility functions; archive v2 examples and utils
