+++ src/examples/v2/example-fetch-nourl.js
@@ -1,15 +1,15 @@
-const ZwiftMemoryMonitor = require('../index.js');
+const ZwiftMemoryMonitor = require('../../index.js');
 
 const zmm = new ZwiftMemoryMonitor(
     {
-    // zwiftapp: 'zwiftapp.exe'
-        log: console.log,
-        type: 'playerprofile'
+        log: console.log
     }
 )
 
-zmm.on('data', (playerProfile) => {
-    console.log(playerProfile)
+console.log('last error:', zmm.lasterror)
+
+zmm.on('data', (playerState) => {
+    console.log(playerState)
 })
 
 zmm.on('status.started', () => {
@@ -30,8 +30,8 @@ zmm.on('status.stopping', () => {
     console.log('status.stopping')
 })
 
-
-zmm.once('ready', () => {
+zmm.on('status.loaded', () => {
+    console.log('status.loaded')
     try {
         zmm.start()
     
@@ -42,3 +42,6 @@ zmm.once('ready', () => {
     }
 })
 
+zmm.once('ready', () => {
+    zmm.loadURL()
+})
