+++ src/examples/v2/profile-search.js
@@ -1,4 +1,4 @@
-const ZwiftMemoryMonitor = require('../index.js');
+const ZwiftMemoryMonitor = require('../../index.js');
 
 const zmm = new ZwiftMemoryMonitor(
   {
