+++ src/util/build-patterns.js
@@ -2,7 +2,7 @@ const fs = require('node:fs')
 const path = require('node:path')
 
 /** @type {Array} */
-const lookup = require('../lookup')
+const patternDefinitions = require('../lookup.js')
 
 const folder = path.resolve(__dirname, '../../build/data')
 console.log(folder)
@@ -11,6 +11,6 @@ if (!fs.existsSync(folder)) {
     fs.mkdirSync(folder,  { recursive: true })
 }
 
-Object.keys(lookup)?.forEach((type) => {
-    fs.writeFileSync(path.resolve(folder, `lookup-${type}.json`), JSON.stringify(lookup[type]), 'utf8')
+Object.keys(patternDefinitions)?.forEach((type) => {
+    fs.writeFileSync(path.resolve(folder, `pattern-${type}.json`), JSON.stringify(patternDefinitions[type]), 'utf8')
 })
