+++ src/utils.js
@@ -1,0 +1,7 @@
+function numberToPattern(number) {
+  return ('00000000' + number.toString(16)).slice(-8).match(/../g).reverse().join(' ');
+}
+
+module.exports = {
+  numberToPattern
+};
